+++ testBalloon-framework-core/src/commonMain/kotlin/de/infix/testBalloon/framework/core/TestCompartment.kt
@@ -69,7 +69,6 @@ public open class TestCompartment(name: String, testConfig: TestConfig) :
          * [testConfig] overrides the compartment's default configuration.
          */
         @Suppress("FunctionName")
-        @TestBalloonExperimentalApi
         public fun MainDispatcher(
             mainDispatcher: CoroutineDispatcher? = null,
             testConfig: TestConfig = TestConfig