+++ CHANGELOG.md
@@ -46,7 +46,11 @@ Migration required:
 
 * Robolectric integration: `robolectricTestSuite` provides a type parameter for the content class instead of a `KClass` parameter. The latter variant is now deprecated.
 * The Gradle integration supports the incubating Gradle JVM Test Suite Plugin and the `com.android.test` plugin out of the box. If one of these plugins is present, tests will be detected in all source sets of the respective Gradle (sub)project. The Gradle property `testBalloon.gradleTestSuiteNamesRegex` is now obsolete and can be deleted.
-* Initial support for the Amper build tool (without IDE support for gutter icons).
+* Experimental support was added for the Amper build tool. IDE gutter icons with Amper builds are currently not supported.
+* Several APIs are now considered stable, removing the need to use the `TestBalloonExperimentalApi` opt-in, notably:
+    * `TestConfig`: permits, `singleThreaded()`, `mainDispatcher()`
+    * `TestCompartment.MainDispatcher()`
+    * Android: `JUnit4RulesContext`
 
 ## 0.8.4 (March 16, 2026)